fixed fill timeseries all values were integers

+++ austaltools/austal_fill_timeseries.py
@@ -1,6 +1,7 @@
 #!/bin/env python3
 
 import argparse
+import numpy as np
 import os
 import logging
 
@@ -160,7 +161,7 @@ def parse_cycle(c_id, c_info, time, dt):
 
     # generate cycle:
     # copy sequence to each start time
-    cycle = pd.Series(0, index=time, name=c_id)
+    cycle = pd.Series(0, index=time, name=c_id, dtype=np.float)
     for x in start:
         for dx, y in sequence.items():
             cycle[x + dx] = y
